requirements.txt
click==6.7
dictmerge==0.2.0
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
python-frontmatter==0.4.2
PyYAML==3.12
six==1.10.0
Werkzeug==0.12.2
